pyproject.toml
[project]
name = "tds-proj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "html-to-markdown>=1.3.3",
    "ipykernel>=6.29.5",
    "requests>=2.32.4",
    "token-count>=0.2.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "openai>=1.28.0",
    "beautifulsoup4>=4.12.3",
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.8.0",
    "markdown>=3.6",
    "semantic-text-splitter>=0.27.0",
    "ollama>=0.2.0",
    "numpy>=1.26.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.0.2",
    "pydantic-core>=2.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*"]
exclude = ["discourse_content*", "course_content*"]
